More bug fixes

+++ exult/trunk/server/objserial.cc
@@ -188,7 +188,7 @@ void Npc_actor_io
 	short& npc_num,
 	short& ident,
 	int& usecode,
-	short *properties,		// Must have room for 12.
+	int *properties,		// Must have room for 12.
 	short& attack_mode,
 	short& alignment,
 	unsigned long& oflags,		// Object flags.
@@ -335,7 +335,7 @@ int Npc_actor_out
 	short npc_num,
 	short ident,
 	int usecode,
-	short properties[12],
+	int properties[12],
 	short attack_mode,
 	short alignment,
 	unsigned long oflags,		// Object flags.
@@ -372,7 +372,7 @@ int Npc_actor_in
 	short& npc_num,
 	short& ident,
 	int& usecode,
-	short properties[12],
+	int properties[12],
 	short& attack_mode,
 	short& alignment,
 	unsigned long& oflags,		// Object flags.
